Create youtube-auto-publish.js
This is a Javascript code for automated video publishing on YouTube Studio. Since maximum allowed listing is 50 videos per page, the maximum automated publishing is capped at 50, beyond that you have to manually click to go to another page.

Disclaimer:
The original Author is not by me, all credits go to the original Author. Dm me if you are the original Author for proper credit.